Server/main/requirements.txt
open3d==0.18.0
numpy==1.26.4
scikit-learn==1.6.0
torch==2.5.1
pillow==11.0.0
opencv-python==4.10.0.84
Flask==3.0.3
ollama==0.4.4
scipy==1.14.1
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1